+++ client/.eslintrc.json
@@ -1,0 +1,37 @@
+{
+  "env": {
+    "browser": true,
+    "es6": true
+  },
+  "extends": ["airbnb", "airbnb/hooks"],
+  "globals": {
+    "Atomics": "readonly",
+    "SharedArrayBuffer": "readonly"
+  },
+  "parser": "babel-eslint",
+  "parserOptions": {
+    "ecmaFeatures": {
+      "jsx": true
+    },
+    "ecmaVersion": 2018,
+    "sourceType": "module"
+  },
+  "plugins": ["react", "prettier"],
+  "rules": {
+    "strict": 0,
+    "react/jsx-one-expression-per-line": 0,
+    "react/jsx-filename-extension": 0,
+    "react/self-closing-comp": 0,
+    "react/button-has-type": 0,
+    "react/prop-types": 1,
+    "react/jsx-props-no-spreading": 0,
+    "prettier/prettier": "error",
+    "max-classes-per-file": 0,
+    "no-console": "off",
+    "no-plusplus": 0,
+    "no-param-reassign": 0,
+    "no-multi-assign": 0,
+    "consistent-return": 0,
+    "no-unused-expressions": 0
+  }
+}
