+++ client/.prettierrc
@@ -1,0 +1,4 @@
+{
+    "singleQuote": true,
+    "trailingComma": "es5"
+}
