+++ package.json
@@ -1,48 +1,25 @@
 {
-  "name": "mern-userlist",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.4.0",
-    "@testing-library/user-event": "^7.2.1",
-    "antd": "^3.26.4",
-    "axios": "^0.19.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-highlight-words": "^0.16.0",
-    "react-hook-form": "^4.4.2",
-    "react-redux": "^7.1.3",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0",
-    "react-search-box": "^2.0.2",
-    "react-search-field": "^1.0.0",
-    "react-search-input": "^0.11.3",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
-    "yup": "^0.28.0"
-  },
+  "name": "mern-userlist-project1",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "server": "npm run server --prefix server",
+    "client": "npm start --prefix client",
+    "start": "concurrently \"npm run server\" \"npm run client\"",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "eslintConfig": {
-    "extends": "react-app"
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mingyangGao1123/mern-userlist-project1.git"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/mingyangGao1123/mern-userlist-project1/issues"
   },
-  "proxy": "http://localhost:3003/"
+  "homepage": "https://github.com/mingyangGao1123/mern-userlist-project1#readme",
+  "devDependencies": {
+    "concurrently": "^5.0.2"
+  }
 }
