+++ server/.eslintrc.json
@@ -1,0 +1,29 @@
+{
+  "env": {
+    "browser": true,
+    "es6": true,
+    "node": true
+  },
+  "extends": ["eslint:recommended"],
+  "globals": {
+    "Atomics": "readonly",
+    "SharedArrayBuffer": "readonly"
+  },
+  "parserOptions": {
+    "ecmaFeatures": {
+      "jsx": true
+    },
+    "ecmaVersion": 2018,
+    "sourceType": "module"
+  },
+  "plugins": ["prettier"],
+  "rules": {
+    "prettier/prettier": "error",
+    "max-classes-per-file": 0,
+    "no-console": "off",
+    "no-plusplus": 0,
+    "no-param-reassign": 0,
+    "no-multi-assign": 0,
+    "consistent-return": 0
+  }
+}
