+++ server/package.json
@@ -22,6 +22,10 @@
   },
   "devDependencies": {
     "concurrently": "^5.0.1",
-    "nodemon": "^2.0.2"
+    "eslint": "^6.8.0",
+    "eslint-plugin-import": "^2.19.1",
+    "eslint-plugin-prettier": "^3.1.2",
+    "nodemon": "^2.0.2",
+    "prettier": "^1.19.1"
   }
 }
